+++ test/test-read_data.R
@@ -1,0 +1,4 @@
+library(readxl)
+# library(openxlsx)
+
+df = read_excel("INPUT/gd_water.xls")